float and double refactorings

+++ app/src/main/java/com/example/moonlight/database/SleepPosition.kt
@@ -14,9 +14,9 @@ data class SleepPosition(
   val sleepPositionTime: Long,
 
   @ColumnInfo(name = "pitch")
-  var pitch: Float,
+  var pitch: Double,
 
   @ColumnInfo(name = "roll")
-  var roll: Float
+  var roll: Double
 
 )
